% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_logging.R
\name{LayoutLogstash}
\alias{LayoutLogstash}
\title{Create the logstash format for the log}
\description{
Refer to: https://github.com/s-fleck/lgr/issues/29#issuecomment-558022635 #' for explanation
}
\section{Super classes}{
\code{\link[lgr:Layout]{lgr::Layout}} -> \code{\link[lgr:LayoutJson]{lgr::LayoutJson}} -> \code{LayoutLogstash}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-format_event}{\code{LayoutLogstash$format_event()}}
\item \href{#method-clone}{\code{LayoutLogstash$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="LayoutJson" data-id="initialize">}\href{../../lgr/html/LayoutJson.html#method-initialize}{\code{lgr::LayoutJson$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="LayoutJson" data-id="set_toJSON_args">}\href{../../lgr/html/LayoutJson.html#method-set_toJSON_args}{\code{lgr::LayoutJson$set_toJSON_args()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="LayoutJson" data-id="toString">}\href{../../lgr/html/LayoutJson.html#method-toString}{\code{lgr::LayoutJson$toString()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_event"></a>}}
\if{latex}{\out{\hypertarget{method-format_event}{}}}
\subsection{Method \code{format_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayoutLogstash$format_event(event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{The logging event that triggers the log creation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayoutLogstash$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
